refactor: Break out type generalization to its own module
